+++ apps/web-cosmos/src/pages/blocks/[height].tsx
@@ -1,6 +1,6 @@
 import withGetStaticProps from '@/pages/withGetStaticProps';
-import type { NextPage } from 'next';
 import BlockDetails from '@/screens/block_details';
+import type { NextPage } from 'next';
 import nextI18NextConfig from '../../../next-i18next.config';
 
 const BlockDetailsPage: NextPage = () => <BlockDetails />;