+++ apps/web-cosmos/src/pages/blocks/index.tsx
@@ -1,6 +1,6 @@
 import withGetStaticProps from '@/pages/withGetStaticProps';
-import type { NextPage } from 'next';
 import Blocks from '@/screens/blocks';
+import type { NextPage } from 'next';
 import nextI18NextConfig from '../../../next-i18next.config';
 
 const BlocksPage: NextPage = () => <Blocks />;