+++ apps/web-cosmos/src/pages/transactions/index.tsx
@@ -1,6 +1,6 @@
 import withGetStaticProps from '@/pages/withGetStaticProps';
-import type { NextPage } from 'next';
 import Transactions from '@/screens/transactions';
+import type { NextPage } from 'next';
 import nextI18NextConfig from '../../../next-i18next.config';
 
 const TransactionsPage: NextPage = () => <Transactions />;