+++ packages/ui/src/components/nav/components/mobile/components/navbar/index.tsx
@@ -8,6 +8,8 @@ import Link from 'next/link';
 import { useRecoilValue } from 'recoil';
 import BigDipperLogoRed from 'shared-utils/assets/big-dipper-red.svg';
 import BigDipperLogoWhite from 'shared-utils/assets/big-dipper-white.svg';
+import WalletDetails from '@/components/nav/components/wallet_details';
+import ConnectWallet from '@/components/nav/components/connect_wallet';
 
 const Navbar = (props: NavbarProps) => {
   const { classes, cx } = useStyles();
@@ -38,6 +40,11 @@ const Navbar = (props: NavbarProps) => {
           <p className="text">{selectedName}</p>
           <ExpandMoreIcon fontSize="small" />
         </div>
+        {/* =================================== */}
+        {/* Wallet Details */}
+        {/* =================================== */}
+        <WalletDetails />
+        <ConnectWallet />
         {/* =================================== */}
         {/* Hamburger */}
         {/* =================================== */}