+++ packages/ui/src/recoil/user/types.ts
@@ -1,0 +1,8 @@
+import { PubKey } from '@/recoil/user/atom';
+
+export interface AtomState {
+  address: string;
+  pubKey: PubKey;
+  walletName: string;
+  loggedIn: boolean;
+}